birding_il_bots/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-secret-manager==2.19.0
google-crc32c==1.5.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
multidict==6.0.5
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
urllib3==2.2.1
yarl==1.9.4
